refactor: terminal routes
